rename integration to e2e
