+++ tox.ini
@@ -1,6 +1,6 @@
 [flake8]
 ignore = W605
-exclude = tests/integration/*
+exclude = tests/e2e/*
 
 per-file-ignores =
     functions/*: E402
